services/product-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-fastapi-instrumentator==6.1.0
psycopg2-binary==2.9.9
pydantic==2.5.0